Lazily `hyperspy.misc.array_tools`
